refactor: rename print to convert since it convert to numbers as well
