+++ components/frontend/default/.test.list
@@ -1,3 +1,3 @@
-- print.mjs
+- convert.mjs
 - payload.mjs
 - index.mjs
